Clean up Moving eips
